+++ packages/medusa/src/index.ts
@@ -1,0 +1,3 @@
+export { default } from "./dataProvider";
+export { default as dataProvider } from "./dataProvider";
+export * from "./authProvider";